MENEX-83: Restructure API Into Modules
