+++ client/src/api/config.js
@@ -4,44 +4,45 @@ export const ENDPOINTS = {
 
     WEBSOCKET: "ws://localhost:3001?user_id=",
 
-    LOGIN: `${API_BASE_URL}/login`,
-    LOGOUT: `${API_BASE_URL}/logout`,
+    LOGIN: `${API_BASE_URL}/api/auth/login`,
+    LOGOUT: `${API_BASE_URL}/api/auth/logout`,
 
-    GET_CURRENT_USER: `${API_BASE_URL}/getCurrentUser`,
-    GET_USER: `${API_BASE_URL}/getUser/:user_id`,
-    GET_PROFILE: `${API_BASE_URL}/getProfile/:handle`,
+    CREATE_USER: `${API_BASE_URL}/api/auth/createUser`,
+    GET_SESSION_USER: `${API_BASE_URL}/api/user/getSessionUser`,
+    GET_USER: `${API_BASE_URL}/api/user/getUserById/:user_id`,
+    GET_PROFILE: `${API_BASE_URL}/api/user/getProfile/:handle`,
 
-    UPDATE_ACCOUNT_SETTINGS: `${API_BASE_URL}/updateAccountSettings`,
-    UPDATE_PROFILE_SETTINGS: `${API_BASE_URL}/updateProfileSettings/:handle`,
+    UPDATE_ACCOUNT_SETTINGS: `${API_BASE_URL}/api/auth/updateAccountSettings`,
+    UPDATE_PROFILE_SETTINGS: `${API_BASE_URL}/api/user/updateProfileSettings/:handle`,
 
-    FOLLOW_USER: `${API_BASE_URL}/followUser`,
-    UNFOLLOW_USER: `${API_BASE_URL}/unfollowUser`,
-    FOLLOW_CHECK: `${API_BASE_URL}/followCheck`,
+    FOLLOW_USER: `${API_BASE_URL}/api/follow/followUser`,
+    UNFOLLOW_USER: `${API_BASE_URL}/api/follow/unfollowUser`,
+    FOLLOW_CHECK: `${API_BASE_URL}/api/follow/followCheck`,
 
-    CREATE_POST: `${API_BASE_URL}/createPost`,
-    UPDATE_POST: `${API_BASE_URL}/updatePost/:postId`,
-    DELETE_POST: `${API_BASE_URL}/deletePost/:postId`,
-    GET_POSTS: `${API_BASE_URL}/getPosts`,
-    GET_USER_POSTS: `${API_BASE_URL}/getUserPosts/:handle`,
+    CREATE_POST: `${API_BASE_URL}/api/post/createPost`,
+    UPDATE_POST: `${API_BASE_URL}/api/post/updatePost/:postId`,
+    DELETE_POST: `${API_BASE_URL}/api/post/deletePost/:postId`,
+    GET_POSTS: `${API_BASE_URL}/api/post/getPosts`,
+    GET_USER_POSTS: `${API_BASE_URL}/api/post/getUserPosts/:handle`,
 
-    CREATE_COMMENT: `${API_BASE_URL}/createComment`,
-    UPDATE_COMMENT: `${API_BASE_URL}/updateComment/:comment_id`,
-    DELETE_COMMENT: `${API_BASE_URL}/deleteComment/:comment_id`,
-    GET_COMMENTS: `${API_BASE_URL}/getComments`,
+    CREATE_COMMENT: `${API_BASE_URL}/api/comment/createComment`,
+    UPDATE_COMMENT: `${API_BASE_URL}/api/comment/updateComment/:comment_id`,
+    DELETE_COMMENT: `${API_BASE_URL}/api/comment/deleteComment/:comment_id`,
+    GET_COMMENTS: `${API_BASE_URL}/api/comment/getComments`,
 
-    GET_CONVERSATIONS: `${API_BASE_URL}/getConversations`,
-    GET_CONVERSATION: `${API_BASE_URL}/getConversation/:conversation_id`,
-    GET_MESSAGES: `${API_BASE_URL}/getMessages`,
-    CREATE_MESSAGE: `${API_BASE_URL}/createMessage`,
-    CREATE_CONVERSATION: `${API_BASE_URL}/createConversation`,
-    UPDATE_CONVERSATION_PARTICIPANTS: `${API_BASE_URL}/updateConversationParticipants`,
-    SET_MESSAGES_AS_READ: `${API_BASE_URL}/setMessagesAsRead`,
+    CREATE_CONVERSATION: `${API_BASE_URL}/api/conversation/createConversation`,
+    UPDATE_CONVERSATION_PARTICIPANTS: `${API_BASE_URL}/api/conversation/updateConversationParticipants`,
+    GET_CONVERSATIONS: `${API_BASE_URL}/api/conversation/getConversations`,
 
-    CREATE_NOTIFICATION: `${API_BASE_URL}/createNotification`,
-    GET_NOTIFICATIONS: `${API_BASE_URL}/getNotifications`,
-    UPDATE_NOTIFICATION: `${API_BASE_URL}/updateNotification`,
+    CREATE_MESSAGE: `${API_BASE_URL}/api/message/createMessage`,
+    SET_MESSAGES_AS_READ: `${API_BASE_URL}/api/message/setMessagesAsRead`,
+    GET_MESSAGES: `${API_BASE_URL}/api/message/getMessages`,
 
-    UPLOAD_PROFILE_PICTURE: `${API_BASE_URL}/settings/uploadProfilePicture`,
+    CREATE_NOTIFICATION: `${API_BASE_URL}/api/notification/createNotification`,
+    SET_NOTIFICATION_AS_READ: `${API_BASE_URL}/api/notification/setNotificationAsRead`,
+    GET_NOTIFICATIONS: `${API_BASE_URL}/api/notification/getNotifications`,
 
-    SEARCH: `${API_BASE_URL}/search`,
+    UPLOAD_PROFILE_PICTURE: `${API_BASE_URL}/api/settings/uploadProfilePicture`,
+
+    SEARCH: `${API_BASE_URL}/api/search/search`,
 };
