+++ server/src/config/db.js
@@ -1,0 +1,24 @@
+const mysql = require('mysql2');
+require('dotenv').config();
+
+//Create mySQL Connection using data stored in .env file
+let meNexus = mysql.createConnection({
+    socketPath  : process.env.DB_SOCKETPATH,
+    host        : process.env.DB_HOST,
+    port        : process.env.DB_PORT,
+    user        : process.env.DB_USER,
+    password    : process.env.DB_PASSWORD,
+    database    : process.env.DB_DATABASE
+});
+
+//Connect to mySQL
+meNexus.connect((err) => {
+    if(err) {
+        console.log(err);
+    } else {
+        console.log("Connected to myNexus!");
+    }
+});
+
+// Export the connection to use it in other modules
+module.exports = meNexus;
